backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
python-multipart
pydantic
loguru
pypdf
reportlab
pymongo
dnspython